+++ packages/design/src/Icon.vue
@@ -6,6 +6,7 @@
 
 <script setup lang="ts">
 import { getConfig } from './config';
+import { IconProps } from './types';
 
 defineOptions({
   name: 'TMIcon',
@@ -13,4 +14,5 @@ defineOptions({
 
 const ui = getConfig('components')?.icon;
 const uiComponent = ui?.component || 'el-icon';
+defineProps<IconProps>();
 </script>